+++ VM/src/manage_zjmp.c
@@ -8,8 +8,7 @@ void    manage_zjmp(t_arena *arena, t_carriage *carr)
     if (carr->carry)
     {
         arg = read_bytes(arena, carr->pc + 1, 2);
-        // ft_printf("from jump |%d|\n", arg);
-        address = (carr->pc + (arg % IDX_MOD)) % MEM_SIZE;
+        address = (unsigned int)(carr->pc + (arg % IDX_MOD)) % MEM_SIZE;
         arena[carr->pc].color_carr = 0;
         carr->pc = address;
         arena[carr->pc].color_carr = carr->color_id;
